feat: add better pipeline state cache serialization feat: update native driver
